+++ src/State/Provider/ParameterProvider.php
@@ -18,7 +18,6 @@ use ApiPlatform\Metadata\Operation;
 use ApiPlatform\State\Exception\ParameterNotSupportedException;
 use ApiPlatform\State\Exception\ProviderNotFoundException;
 use ApiPlatform\State\ParameterNotFound;
-use ApiPlatform\State\ParameterProviderInterface;
 use ApiPlatform\State\ProviderInterface;
 use ApiPlatform\State\Util\ParameterParserTrait;
 use ApiPlatform\State\Util\RequestParser;
@@ -99,13 +98,15 @@ final class ParameterProvider implements ProviderInterface
                 continue;
             }
 
-            if (!\is_string($provider) || !$this->locator->has($provider)) {
-                throw new ProviderNotFoundException(\sprintf('Provider "%s" not found on operation "%s"', $provider, $operation->getName()));
+            if (\is_string($provider)) {
+                if (!$this->locator->has($provider)) {
+                    throw new ProviderNotFoundException(\sprintf('Provider "%s" not found on operation "%s"', $provider, $operation->getName()));
+                }
+
+                $provider = $this->locator->get($provider);
             }
 
-            /** @var ParameterProviderInterface $providerInstance */
-            $providerInstance = $this->locator->get($provider);
-            if (($op = $providerInstance->provide($parameter, $values, $context)) instanceof Operation) {
+            if (($op = $provider->provide($parameter, $values, $context)) instanceof Operation) {
                 $operation = $op;
             }
         }